+++ packages/icons/src/__generated__/nav/data/names.ts
@@ -122,6 +122,8 @@ const names: NavIconName[] = [
   'helpCenterQuestionMark',
   'cluster',
   'borrowProduct',
+  'coinbaseOneLogo',
+  'qrCode',
 ];
 
 export default names;