+++ packages/illustrations/src/__generated__/pictogram/svgJs/dark/lightningNetworkSend-0.js
@@ -1,0 +1,3 @@
+module.exports = {
+  content: `<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 48 48"><g clip-path="url(#a)"><path fill="#45DAF6" d="M40 20H0v16h40V20Z"/><path fill="#ECD069" d="M40 36c4.421 0 8-3.579 8-8 0-4.421-3.579-8-8-8-4.421 0-8 3.579-8 8 0 4.421 3.579 8 8 8Z"/><path fill="#464B55" d="M32 14.77H20.93L23.53 0H8.76L4.2 25.84h9.23L9.53 48 32 14.77Z"/><path fill="#588AF5" d="m17.64 36 10.82-16H5.23L4.2 25.84h9.23L11.64 36h6Z"/></g><defs><clipPath id="a"><path fill="#FFFFFF" d="M0 0h48v48H0z"/></clipPath></defs></svg>`,
+};